docker/models-requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Computer Vision and Image Processing
opencv-python==4.8.1.78
Pillow==10.0.1
numpy>=1.26.0,<2.2.0

# Machine Learning and AI
tensorflow-cpu>=2.13.0,<2.16.0
scikit-learn==1.3.0
scipy==1.11.1
scikit-image==0.21.0
mahotas==1.4.13
ultralytics==8.0.196

# Data processing and utilities
pandas==2.0.3

# Additional utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Development and testing
pytest==7.4.2
pytest-asyncio==0.21.1